% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_module.R
\name{use_module}
\alias{use_module}
\title{Install course materials}
\usage{
use_module(module, destdir = NULL)
}
\arguments{
\item{module}{The name of the module}

\item{destdir}{The path on your computer where you would like the course
installed. If \code{NULL}, defaults to your Desktop or some other conspicuous
place.}
}
\description{
\code{use_module()} will install the materials for a given module in the course
Statistical Programming in R on your computer. Then, it will open a new
RStudio Project containing the files you'll need for that module.
}
